Clean up examples dir a little
